# musicbox-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API",
        "description": "API documentation for MusicBox project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get all albums (with optional filter by year)",
                "operationId": "735609843dcfb89ee4e12e340eddca25",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter by album release year",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Albums retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Create a new album",
                "operationId": "2252aa8fd63b7d2fa1203c1735943d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "year",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "artist_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get album by ID",
                "operationId": "71e9c4179582355e89419e6f6994c418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album retrieved successfully"
                    },
                    "404": {
                        "description": "Album not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Update an album",
                "operationId": "1300ec85cc590b1ae1468c9bf8db122d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "artist_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Delete an album",
                "operationId": "cabfc7c152115607d310827d46a82d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Use token from login",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}